Add git submodules support

+++ polemarch/main/tests/executions.py
@@ -636,8 +636,12 @@ class ProjectTestCase(BaseExecutionsTestCase):
     def tearDown(self):
         super(ProjectTestCase, self).tearDown()
         repo_dir = getattr(self, 'repo_dir', None)
+        submodule_dir = getattr(self, 'submodule_dir', None)
         if repo_dir:
-            shutil.rmtree(repo_dir)
+            if os.path.exists(repo_dir):
+                shutil.rmtree(repo_dir)
+            if os.path.exists(submodule_dir):
+                shutil.rmtree(submodule_dir)
 
     def wip_manual(self, project_data):
         files = self.generate_playbook(self.get_project_dir(**project_data))
@@ -733,12 +737,13 @@ class ProjectTestCase(BaseExecutionsTestCase):
         proj = self.get_model_filter('Project', pk=project_data['id']).get()
         proj_config = getattr(proj, 'config', None)
         self.assertTrue(proj_config is not None)
-        # Check modules
+        # Check modules and git submodules sync
         proj_modules = proj.module.filter(path__startswith='polemarch.project.')
         self.assertEqual(proj_modules.count(), 1)
         proj_module = proj_modules.first()
         self.assertEqual(proj_module.name, 'test_module')
         self.assertEqual(proj_module.data['short_description'], 'Test module')
+        self.assertTrue(os.path.exists(os.path.join(proj.path, 'sm1', 'test_module.py')))
         return dict(playbook_count=len(self.revisions), execute=True)
 
     def make_test_templates(self, project_data):
@@ -1129,17 +1134,25 @@ class ProjectTestCase(BaseExecutionsTestCase):
         pm_yaml['view'] = test_yaml_view
         # Prepare repo
         self.repo_dir = tempfile.mkdtemp()
+        self.submodule_dir = "{}_submodule".format(self.repo_dir)
         self.generate_playbook(self.repo_dir, ['main.yml'])
         self.generate_playbook(self.repo_dir, ['.polemarch.yaml'], data=dump(pm_yaml))
         self.generate_playbook(self.repo_dir, ['ansible.cfg'], data=test_ansible_cfg)
-        lib_dir = self.repo_dir + '/lib'
+        lib_dir = self.submodule_dir
         if not os.path.exists(lib_dir):
             os.makedirs(lib_dir)
             self.generate_playbook(lib_dir, ['test_module.py'], data=test_module_content)
+        # Create submodule
+        submodule_repo = git.Repo.init(self.submodule_dir)
+        submodule_repo.index.add(['test_module.py'])
+        submodule_repo.index.commit("no message")
         repo = git.Repo.init(self.repo_dir)
-        repo.index.add([
-            "main.yml", ".polemarch.yaml", "ansible.cfg", lib_dir + '/test_module.py'
-        ])
+        # Add submodules
+        repo.git.submodule(
+            'add', '../{}/.git'.format(self.submodule_dir.split('/')[-1]), 'lib'
+        )
+        repo.git.submodule('add', '{}/.git'.format(self.submodule_dir), 'sm1')
+        repo.index.add(["main.yml", ".polemarch.yaml", "ansible.cfg"])
         repo.index.commit("no message")
         first_revision = repo.head.object.hexsha
         repo.create_head('new_branch')
